Docs(packages): table column demo
